Load Dropzone just on sales
Load dropzone plugin just in sales view.

By default the project always includes the js Dropzone plugin,
and the web explorer was always calling it,
this always raises an error from element not found in views not needed.

To solve this, just add an if with the id from dropzone,
this avoids the automatic load and just loads when the element is needed.

+++ app/assets/javascripts/pages/sales.js
@@ -2,49 +2,51 @@ Dropzone.autoDiscover = false
 datatransfer = new DataTransfer
 
 $(document).ready(function () {
-  var dropzone = new Dropzone('#sales-dropzone', {
-    addRemoveLinks: true,
-    acceptedFiles: '.png,.jpg,.jpeg',
-    url: '/file/post',
-    autoProcessQueue: false,
-    uploadMultiple: true,
-    parallelUploads: 5,
-    maxFilesize: 5,
-    filesizeBase: 1000,
-  })
+  if (document.getElementById('sales-dropzone')) {
+    var dropzone = new Dropzone('#sales-dropzone', {
+      addRemoveLinks: true,
+      acceptedFiles: '.png,.jpg,.jpeg',
+      url: '/file/post',
+      autoProcessQueue: false,
+      uploadMultiple: true,
+      parallelUploads: 5,
+      maxFilesize: 5,
+      filesizeBase: 1000,
+    })
 
-  dropzone.on('removedfile', function() {
-    datatransfer.items.remove(datatransfer.items.length - 1)
-  })
+    dropzone.on('removedfile', function() {
+      datatransfer.items.remove(datatransfer.items.length - 1)
+    })
 
-  dropzone.on('addedfile', function(file) {
-    datatransfer.items.add(file)
-    if (validImageType(file) && validItemsLength()) {
-      addFile(file)
-    } else {
-      dropzone.removeFile(file)
+    dropzone.on('addedfile', function(file) {
+      datatransfer.items.add(file)
+      if (validImageType(file) && validItemsLength()) {
+        addFile(file)
+      } else {
+        dropzone.removeFile(file)
+      }
+    })
+
+    function validImageType(file) {
+      return (file.type === 'image/jpeg' || file.type === 'image/png')
     }
-  })
 
-  function validImageType(file) {
-    return (file.type === 'image/jpeg' || file.type === 'image/png')
-  }
-
-  function validItemsLength() {
-    return (datatransfer.items.length < 6)
-  }
-
-  function addFile() {
-    $('.info-dragdrop')[0].classList.add('d-none')
-    $('#fileMultipleItems')[0].files = datatransfer.files
-  }
-
-  $('#new_sale').submit(function(e) {
-    if (datatransfer.items.length === 0){
-      e.preventDefault()
-      $(window).scrollTop(0)
-      $('#image-error').text('Por favor suba una foto.')
-      return false
+    function validItemsLength() {
+      return (datatransfer.items.length < 6)
     }
-  })
+
+    function addFile() {
+      $('.info-dragdrop')[0].classList.add('d-none')
+      $('#fileMultipleItems')[0].files = datatransfer.files
+    }
+
+    $('#new_sale').submit(function(e) {
+      if (datatransfer.items.length === 0){
+        e.preventDefault()
+        $(window).scrollTop(0)
+        $('#image-error').text('Por favor suba una foto.')
+        return false
+      }
+    })
+  }
 })
